BORM-30158 Pointline: Fixed computation of bounding boxes.
The bounding boxes must consider the volumetric points as well (like circles, rectangles, spheres, boxes) to compute properly the bounding boxes used by kd-trees. The previous implementation took only a single point (x, y, z) and merge it with the currently computed bounding box which was incorrected in the case of e.g. rectangles as it is used by Pointline.